Break parseValue up into more files
